The LogIn component is created on the desktop.
In the backend file UserController.js, the code for LogIn is added.

In the backend file, routes.js, the route Post / user / login / is added.

CORS error is fixed, when editing /config/security.js, allRoutes = true.

Fixed the User model so that the table is created with the name users and not user.

The generateAuthToken.js helper is created

/Api/policies/auth.js is created